requirements.txt
aiofiles==24.1.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
black==24.8.0
certifi==2024.7.4
click==8.1.7
ecdsa==0.19.0
fastapi==0.112.1
fastapi-cache2==0.2.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.2.2
pyasn1==0.6.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
redis==5.0.8
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqladmin==0.18.0
SQLAlchemy==2.0.32
starlette
time-machine==2.15.0
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
WTForms==3.1.2
fastapi-debug-toolbar
gunicorn